hiver_data_commons::optimistic_lock

Trait Version

pub trait Version {
    // Required methods
    fn version(&self) -> i64;
    fn set_version(&mut self, version: i64);

    // Provided method
    fn increment_version(&mut self) { ... }
}

Trait for versioned entities supporting optimistic locking. 支持乐观锁的版本化实体 trait。

Entities implementing this trait carry a monotonically increasing version number that is checked on update to prevent lost updates. 实现此 trait 的实体携带一个单调递增的版本号， 在更新时进行检查以防止更新丢失。

This is the Rust equivalent of Spring Data’s @Version annotation. 这是 Spring Data @Version 注解的 Rust 等价物。

Example / 示例

use hiver_data_commons::optimistic_lock::Version;

#[derive(Debug, Clone)]
struct VersionedUser
{
    id: i32,
    name: String,
    version: i64,
}

impl Version for VersionedUser
{
    fn version(&self) -> i64
    {
        self.version
    }

    fn set_version(&mut self, version: i64)
    {
        self.version = version;
    }
}

Required Methods

fn version(&self) -> i64

Get the current version number. 获取当前版本号。

fn set_version(&mut self, version: i64)

Set the version number. 设置版本号。

Provided Methods

fn increment_version(&mut self)

Increment the version number by one. 将版本号递增一。