hiver_data_commons::specification

Struct SpecFactories

pub struct SpecFactories;

Factory for creating common specifications. 创建常用 specification 的工厂。

Since Rust does not allow inherent impl blocks on traits, these factory methods are provided on a separate struct. 由于 Rust 不允许在 trait 上使用固有 impl 块， 这些工厂方法在单独的结构体上提供。

Example / 示例

use hiver_data_commons::specification::{SpecFactories, SpecValue, Specification};

let spec = SpecFactories::eq("name", SpecValue::String("Alice".into()));
assert!(spec.to_predicate().is_some());

Implementations

impl SpecFactories

pub fn always() -> AlwaysSpec

Create a specification that always matches (no constraints). 创建总是匹配的 specification（无约束）。

pub fn eq(field: impl Into<String>, value: SpecValue) -> SimpleSpec

Create an equality specification. 创建相等 specification。

pub fn ne(field: impl Into<String>, value: SpecValue) -> SimpleSpec

Create an inequality specification. 创建不等 specification。

pub fn gt(field: impl Into<String>, value: SpecValue) -> SimpleSpec

Create a greater-than specification. 创建大于 specification。

pub fn lt(field: impl Into<String>, value: SpecValue) -> SimpleSpec

Create a less-than specification. 创建小于 specification。

pub fn gte(field: impl Into<String>, value: SpecValue) -> SimpleSpec

Create a greater-than-or-equal specification. 创建大于等于 specification。

pub fn lte(field: impl Into<String>, value: SpecValue) -> SimpleSpec

Create a less-than-or-equal specification. 创建小于等于 specification。

pub fn like(field: impl Into<String>, pattern: impl Into<String>) -> SimpleSpec

Create a LIKE specification. 创建 LIKE specification。

pub fn in_pred(field: impl Into<String>, values: Vec<SpecValue>) -> SimpleSpec

Create an IN specification. 创建 IN specification。

pub fn not_in(field: impl Into<String>, values: Vec<SpecValue>) -> SimpleSpec

Create a NOT IN specification. 创建 NOT IN specification。

pub fn is_null(field: impl Into<String>) -> SimpleSpec

Create an IS NULL specification. 创建 IS NULL specification。

pub fn is_not_null(field: impl Into<String>) -> SimpleSpec

Create an IS NOT NULL specification. 创建 IS NOT NULL specification。

pub fn between( field: impl Into<String>, low: SpecValue, high: SpecValue, ) -> SimpleSpec

Create a BETWEEN specification. 创建 BETWEEN specification。

pub fn and_spec( s1: impl Specification + 'static, s2: impl Specification + 'static, ) -> AndSpec

Combine two specifications with AND. 使用 AND 组合两个 specification。

pub fn or_spec( s1: impl Specification + 'static, s2: impl Specification + 'static, ) -> OrSpec

Combine two specifications with OR. 使用 OR 组合两个 specification。

pub fn not_spec(s: impl Specification + 'static) -> NotSpec

Negate a specification. 取反一个 specification。